requirements.txt
Django==3.2
djangorestframework==3.12.4
mysqlclient==2.0.3
djangorestframework-simplejwt>=4.9.0
